backend/requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
Flask-Session==0.8.0
Flask-Mail==0.10.0
Werkzeug==3.0.3
pymongo==4.8.0
redis==5.0.8
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
Authlib==1.2.1
PyJWT==2.8.0
pyotp==2.9.0
qrcode[pil]==7.4.2
PyPDF2==3.0.1
pdfplumber==0.11.0
pdfminer.six==20231228
openai==1.40.6
anthropic==0.34.2
google-generativeai==0.8.3
python-dotenv==1.0.1
python-slugify==8.0.4
email-validator==2.2.0
requests==2.32.3
certifi==2024.8.30
charset-normalizer==3.3.2
text-unidecode==1.3
stripe==8.8.0
async-timeout==4.0.3
urllib3==2.2.2
dnspython==2.6.1
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
colorama==0.4.6
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
Pillow==10.4.0
pydantic==2.5.3